Implement persisttent configuration
We don't want to have to set the --url, --tenant, and --token options
every single time we issue an okapi command.

This introduces the concept of configuration very similar to the
heroku command line with a configuration file that is tied closely to
environment vairables.

```
$ okapi config
{
  "OKAPI_URL": "https://okapi.frontside.io",
}

$ okapi config:set OKAPI_TENANT=fs
$ okapi config
{
  "OKAPI_URL": "https://okapi.frontside.io",
  "OKAPI_TENANT": "fs"
}

$ okapi config:delete OKAPI_TENANT
$ okapi config
{
  "OKAPI_URL": "https://okapi.frontside.io"
}
```

By default, configuration settings are stored in $HOME/.okapi however,
you can pass the `--config` option to specify a separate config file.

```
$ okapi --config /path/to/my/config modules:index
```