Display default configuration file location.
Clamp uses the 'inspect' method to display a default argument to the
user. However, in our case, when this is an object, it makes the
output look like `#<Okapi::ImplicitConfig:0x00007fc7849108e0>` This
makes the config point out the location of its config. so that it will
be something like `/Users/cowboyd/.okapi`