Make better interface for instantiating a client.
Instantiating a client is the most important thing you'll want to do
with the okapi library, so why do you need to create an object to
create the object?

Instead of a spring-loaded interface, this just takes them straight up
in the constructor, and then lazily constructs the settings object
as-needed.