Delete and Post from the CLI
In the beginning, there was only `get`, `tenant:get`, and `user:get`,
which was great for readonly operations, but we need, in our
interaction with mod-configuration to delete all the configuration
entries and then create them again.

This adds two new commands to the CLI (and the ruby client): `create`
to issue POST requests. You can pass the body of the newly minted
resource via stdin:

```
$ echo '{"some": "json"}' | okapi create /path/to/some/resource
```

By the same token, you can issue DELETE requests from the client, and
from the command line as well:

```
$ okapi destroy /path/to/some/resource
```