Let rubygems find the `okapi` executable.
Out of the box, bundler creates the gemspec with the `./exe` as a
place to look for executables so that it can install binstubs into
your path. I suppose this change was made since technically these
scripts like `./bin/okapi` aren't truly "binary" files in the sense of
the word that was in place when most of the executables in the system
were in fact compiled binaries.

As a result, when you install the `okapi` gem from rubygems, the cli at
`bin/okapi` is not added to the path automatically.

However, the convention is to put executable stuff in a directory
called `bin` even if it isn't binary, so we'll stick with that. and
instead of telling rubygems to look for scripts in `./exe`, we tell it
to go ahead and look in `./bin`