Demonstrate that CLI prompts for input even without an OKAPI URL
If you don't have a URL configured for your OKAPI cluster, and you
don't have a tenant configured either, then logging in is a
non-starter, and prompting for input is pointless.

This failing test demonstrates that the CLI does in fact attempt to
read console input from the user even when we know that it is
impossible to do anything with them.