Prefetch url and tenant configuration params before logging in.
If you don't have a URL configured for your OKAPI cluster, and you
don't have a tenant configured either, then logging in is a
non-starter, and prompting for input is pointless.

This accesses the `url` and `tenant` configuration parameters off of
the `Okapi::Settings` object before the login process. These methods
throw exceptions if the parameter is not present, so it will stop the
process before bothering the user to enter in their username and
password.

Also, the wording of the errors are updated to indicate that you can
use environment variables, command line options OR persistent
configuration variables to set your url and tenant.