Update okapi docs to reflect preferred API
When we started the okapi gem, the idea was that we would have methods
for the semantic endpoints contained in an okapi cluster. I.e. to look
up the modules installed on a cluster, we would call `client.modules`,
or alternatively `okapi modules:index` from the command line.

Since however, it's made life much simpler, and the client less
brittle to support the simple RESTful verbs `get`, `put`, `post`,
`delete`, since those won't be changing any time soon.

This removes those old interfaces and upgrades the docs to reflect the
new and preferred usage. Because this is a potentially breaking
change, the major version number has also been increased.

As a bonus, a shiny new PR template can ship with 2.0!